Styles page to show that post has been submitted

+++ app/controllers/journal_entries_controller.rb
@@ -8,17 +8,22 @@ class JournalEntriesController < ApplicationController
   end
 
   def create
-    binding.pry
     user_id = current_user.id
-    new_entry = google_service.create_file(journal_params[:tag])
-    `touch entry#{user_id}.txt`
-    `echo "#{journal_params[:body]}" >> entry#{user_id}.txt`
-    reply = google_service.drive.insert_file(new_entry,
-                                             upload_source: "entry#{user_id}.txt",
-                                             content_type: 'text/plain')
-    `rm entry#{user_id}.txt`
-    @entry = JournalEntry.create(file_id: reply.id, user: current_user)
-    redirect_to journal_entry_path(@entry)
+    if journal_params["user_id"].to_i == current_user.id
+      new_entry = google_service.create_file(params[:tag])
+      `touch entry#{user_id}.txt`
+      `echo "#{params[:body]}" >> entry#{user_id}.txt`
+      reply = google_service.drive.insert_file(new_entry,
+        upload_source: "entry#{user_id}.txt",
+        content_type: 'text/plain')
+      `rm entry#{user_id}.txt`
+      @entry = JournalEntry.create(file_id: reply.id)
+      render json: {created: "success"}
+      # redirect_to journal_entry_path(@entry)
+    else
+      render json: {created: "error"}
+
+    end
   end
 
   def show
@@ -32,6 +37,7 @@ class JournalEntriesController < ApplicationController
   end
 
   def journal_params
-    params.require(:journal_entry).permit(:tag, :body)
+    params.permit!
+    params.except("controller","action","tag","body")
   end
 end
